+++ src/components/user/index.ts
@@ -12,3 +12,4 @@ export * from "./addresses";
 export * from "./settings";
 export * from "./profile";
 export * from "./notifications";
+export * from "./orders";